requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Pillow==8.1.0
pony==0.7.14
requests==2.25.1
six==1.15.0
urllib3==1.26.3
vk-api==11.9.1
